+++ src/picker/typer/ydotool.py
@@ -3,6 +3,35 @@ from subprocess import run
 from ..abstractionhelper import is_installed
 from .typer import Typer
 
+# From /usr/include/linux/input-event-codes.h
+_KEYCODES = {
+    "0": "11",
+    "1": "2",
+    "2": "3",
+    "3": "4",
+    "4": "5",
+    "5": "6",
+    "6": "7",
+    "7": "8",
+    "8": "9",
+    "9": "10",
+    "a": "30",
+    "b": "48",
+    "c": "46",
+    "d": "32",
+    "e": "18",
+    "f": "33",
+}
+
+_PRESS = "1"
+_RELEASE = "0"
+
+_INSERT_KEY_CODE = "110"
+_LEFT_CTRL_KEY_CODE = "29"
+_LEFT_SHIFT_KEY_CODE = "42"
+_SPACE_KEY_CODE = "57"
+_U_KEY_CODE = "22"
+
 
 class YdotoolTyper(Typer):
     @staticmethod
@@ -22,16 +51,22 @@ class YdotoolTyper(Typer):
     def type_numerical(self, codepoints: list[int], active_window: str) -> None:
         keypresses = []
         for codepoint in codepoints:
-            keypresses.append = [
-                (self.__get_event_code("LeftCtrl") + ":1"),
-                (self.__get_event_code("LeftShift") + ":1"),
-                (self.__get_event_code("u") + ":1"),
-                (self.__get_event_code("u") + ":0"),
-                (self.__get_event_code("LeftShift") + ":0"),
-                (self.__get_event_code("LeftCtrl") + ":0"),
-                (self.__get_event_code(f"{codepoint:x}") + ":1"),
-                (self.__get_event_code(f"{codepoint:x}") + ":0"),
-            ]
+            keypresses.extend(
+                [
+                    f"{_LEFT_CTRL_KEY_CODE}:{_PRESS}",
+                    f"{_LEFT_SHIFT_KEY_CODE}:{_PRESS}",
+                    f"{_U_KEY_CODE}:{_PRESS}",
+                    f"{_U_KEY_CODE}:{_RELEASE}",
+                    f"{_LEFT_SHIFT_KEY_CODE}:{_RELEASE}",
+                    f"{_LEFT_CTRL_KEY_CODE}:{_RELEASE}",
+                ]
+            )
+
+            keypresses.extend(
+                f"{_KEYCODES[digit]}:{action}" for digit in f"{codepoint:x}" for action in (_PRESS, _RELEASE)
+            )
+
+            keypresses.extend([f"{_SPACE_KEY_CODE}:{_PRESS}", f"{_SPACE_KEY_CODE}:{_RELEASE}"])
 
         run(["ydotool", "key", *keypresses])
 
@@ -40,39 +75,9 @@ class YdotoolTyper(Typer):
             [
                 "ydotool",
                 "key",
-                (self.__get_event_code("LeftShift") + ":1"),
-                (self.__get_event_code("Insert") + ":1"),
-                (self.__get_event_code("Insert") + ":0"),
-                (self.__get_event_code("LeftShift") + ":0"),
+                f"{_LEFT_SHIFT_KEY_CODE}:{_PRESS}",
+                f"{_INSERT_KEY_CODE}:{_PRESS}",
+                f"{_INSERT_KEY_CODE}:{_RELEASE}",
+                f"{_LEFT_SHIFT_KEY_CODE}:{_RELEASE}",
             ]
         )
-
-    def __get_event_code(self, char: str) -> str:
-        return str(self._keycodes[char])
-
-    # From /usr/include/linux/input-event-codes.h
-    _keycodes = {
-        "0": 11,
-        "1": 2,
-        "2": 3,
-        "3": 4,
-        "4": 5,
-        "5": 6,
-        "6": 7,
-        "7": 8,
-        "8": 9,
-        "9": 10,
-        "a": 30,
-        "b": 48,
-        "c": 46,
-        "d": 32,
-        "e": 18,
-        "f": 33,
-        "u": 22,
-        "Enter": 28,
-        "LeftCtrl": 29,
-        "LeftShift": 42,
-        "BackSpace": 14,
-        "Tab": 15,
-        "Insert": 110,
-    }